Sorting Visualizer
Is sorting visualizer a good project?
Image result for sorting visualizer project
This project is a good start for beginners and a refresher for professionals who have dabbled in data structures and algorithms using Javascript before and also web developers. The methodology can be applied to showcase any algorithm of one's choosing, so feel free to innovate!